pyproject.toml
[project]
name = "gedion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.0.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "wikipedia>=1.4.0",
]
